# perf: Improve diff performance