--- a/src/client/components/DiffEditor.vue
+++ b/src/client/components/DiffEditor.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { nextTick, onMounted, ref, toRefs, watchEffect } from 'vue'
-import { diff_match_patch as Diff } from 'diff-match-patch'
 import { useCodeMirror } from '../logic/codemirror'
 import { guessMode } from '../logic/utils'
 import { enableDiff, lineWrapping } from '../logic/state'
+import { calucateDiffWithWorker } from '../worker/diff'
 
 const props = defineProps<{ from: string; to: string }>()
 const { from, to } = toRefs(props)
@@ -49,58 +49,58 @@ onMounted(() => {
 
     await nextTick()
 
-    cm1.operation(() => {
-      cm2.operation(() => {
-        // clean up marks
-        cm1.getAllMarks().forEach(i => i.clear())
-        cm2.getAllMarks().forEach(i => i.clear())
-        new Array(cm1.lineCount() + 2)
-          .fill(null!)
-          .map((_, i) => cm1.removeLineClass(i, 'background', 'diff-removed'))
-        new Array(cm2.lineCount() + 2)
-          .fill(null!)
-          .map((_, i) => cm2.removeLineClass(i, 'background', 'diff-added'))
+    cm1.startOperation()
+    cm2.startOperation()
 
-        if (showDiff) {
-          const diff = new Diff()
-          const changes = diff.diff_main(l, r)
-          diff.diff_cleanupSemantic(changes)
+    // clean up marks
+    cm1.getAllMarks().forEach(i => i.clear())
+    cm2.getAllMarks().forEach(i => i.clear())
+    new Array(cm1.lineCount() + 2)
+      .fill(null!)
+      .map((_, i) => cm1.removeLineClass(i, 'background', 'diff-removed'))
+    new Array(cm2.lineCount() + 2)
+      .fill(null!)
+      .map((_, i) => cm2.removeLineClass(i, 'background', 'diff-added'))
 
-          const addedLines = new Set()
-          const removedLines = new Set()
+    if (showDiff) {
+      const changes = await calucateDiffWithWorker(l, r)
 
-          let indexL = 0
-          let indexR = 0
-          changes.forEach(([type, change]) => {
-            if (type === 1) {
-              const start = cm2.posFromIndex(indexR)
-              indexR += change.length
-              const end = cm2.posFromIndex(indexR)
-              cm2.markText(start, end, { className: 'diff-added-inline' })
-              for (let i = start.line; i <= end.line; i++) addedLines.add(i)
-            }
-            else if (type === -1) {
-              const start = cm1.posFromIndex(indexL)
-              indexL += change.length
-              const end = cm1.posFromIndex(indexL)
-              cm1.markText(start, end, { className: 'diff-removed-inline' })
-              for (let i = start.line; i <= end.line; i++) removedLines.add(i)
-            }
-            else {
-              indexL += change.length
-              indexR += change.length
-            }
-          })
+      const addedLines = new Set()
+      const removedLines = new Set()
 
-          Array.from(removedLines).forEach(i =>
-            cm1.addLineClass(i, 'background', 'diff-removed'),
-          )
-          Array.from(addedLines).forEach(i =>
-            cm2.addLineClass(i, 'background', 'diff-added'),
-          )
+      let indexL = 0
+      let indexR = 0
+      changes.forEach(([type, change]) => {
+        if (type === 1) {
+          const start = cm2.posFromIndex(indexR)
+          indexR += change.length
+          const end = cm2.posFromIndex(indexR)
+          cm2.markText(start, end, { className: 'diff-added-inline' })
+          for (let i = start.line; i <= end.line; i++) addedLines.add(i)
+        }
+        else if (type === -1) {
+          const start = cm1.posFromIndex(indexL)
+          indexL += change.length
+          const end = cm1.posFromIndex(indexL)
+          cm1.markText(start, end, { className: 'diff-removed-inline' })
+          for (let i = start.line; i <= end.line; i++) removedLines.add(i)
+        }
+        else {
+          indexL += change.length
+          indexR += change.length
         }
       })
-    })
+
+      Array.from(removedLines).forEach(i =>
+        cm1.addLineClass(i, 'background', 'diff-removed'),
+      )
+      Array.from(addedLines).forEach(i =>
+        cm2.addLineClass(i, 'background', 'diff-added'),
+      )
+    }
+
+    cm1.endOperation()
+    cm2.endOperation()
   })
 })
 </script>