This PR makes diff faster for about 4.5x.

before: 1911ms,1836ms,1765ms
after: 374ms,464ms,385ms

refs #1